fe: Move resource filter and sort functions to separate util functions
